Refactor card creation function into it's own file
Given how large just one function is it's easy to see how it's going to
bloat out the main.ts. I don't think having a single queries.ts file is
going to work long term either, but at least it's easier to refactor
when it's pulled out of the electron code.