Add tests for the createCard query
Also updating the code to allow for a test db to be used. To ensure that
the test db is set up properly I've added a testSetup script that should
be run before the test suite. Ideally, I could add it into vitest, but
from what I can tell vitest doesn't have an ability to run a file once
on start up, so that might need to be done at the shell level.