Allow deletion of the card through the right click menu
This uses the electron APIs to build and display a specific context menu
when right clicking on a card row. Because I can't add any additional
data to the menu item itself, I need to pass in all the information when
building the entire menu. I don't see this being an issue for
performance since all the actions should be tied to the individual card,
and that card can be uniquely identified by it's id, so I shouldn't need
to pass in much more than that.